\name{utils}
\alias{profileDataCycles}
\title{Profile Data Utilities}
\description{
  Utility functions for developing new profile data analysis tools.
}
\usage{
  profileDataCycles(pd, GC)
}
\arguments{
  \item{pd}{profile data as returned by \code{readProfileData}.}
  \item{GC}{logical; include GC information or not.}
}
\details{
   These functions are experimental and subject to change.
}
\author{ Luke Tierney }
\keyword{programming}
\keyword{utilities}
