\name{rgl-package}
\title{Tools for examining R profile output}
\alias{proftools-package}
\alias{proftools}
\description{
  Tools for examining and displaying ouptut from the \code{Rprof} R
  profiling tool.
}
\details{
  \pkg{proftools} is
  %%See the first example below to display the ChangeLog.
}
%%\seealso{****}
%%\examples{
%%file.show(system.file("NEWS", package = "proftools"))
%%example(hotPaths)
%%}
\keyword{programming}
\keyword{utilities}
